requirements.txt
# Economic Terminal Dependencies
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# Configuration
python-dotenv>=1.0.0

# Database
SQLAlchemy>=2.0.25
psycopg[binary]>=3.1.0
alembic>=1.13.0

# HTTP/API Clients
aiohttp>=3.9.0
yfinance>=0.2.35
fredapi>=0.5.1

# Scheduling
APScheduler>=3.10.0

# Email
sendgrid>=6.11.0

# Data Processing
feedparser>=6.0.10
requests>=2.31.0
pydantic>=2.0.0

# Utilities
loguru>=0.7.0
pytz>=2023.3